+++ pipeline/__init__.py
@@ -4,6 +4,8 @@ from .mapper import StructuralMapper, MappingError
 from .pruner import SurgicalPruner, NOISE_SELECTORS
 from .qa import ZeroNoiseAssertionGate, AssertionError_
 from .transform import MarkdownConverter, BoundaryWrapper
+from .scorer import ContentAwareScorer, ScoredBlock
+from .validator import ContentValidator, ValidationResult
 
 __all__ = [
     "StructuralMapper",
@@ -14,4 +16,8 @@ __all__ = [
     "AssertionError_",
     "MarkdownConverter",
     "BoundaryWrapper",
+    "ContentAwareScorer",
+    "ScoredBlock",
+    "ContentValidator",
+    "ValidationResult",
 ]